Challenge: Loopy language
https://www.khanacademy.org/computing/computer-programming/html-js-jquery/jquery-tools-techniques/pc/challenge-loopy-language

Challenge: Loopy language
Translate the paragraphs
This webpage is about Pig Latin, a fun language. You'll use jQuery to translate each of the paragraphs to Pig Latin. Using either a for loop or the each() method, find each paragraph, send its text through the toPigLatin function, and set the paragraph to that text.